add test cases, factory

+++ app/src/main/java/org/example/mockProvider/Stock.java
@@ -1,10 +1,12 @@
 package org.example.mockProvider;
 
+import java.math.BigDecimal;
+
 public class Stock {
     private String ticker;
-    private Double curr_price;
+    private BigDecimal curr_price;
 
-    public Stock(String ticker, Double curr_price){
+    public Stock(String ticker, BigDecimal curr_price){
         this.ticker = ticker;
         this.curr_price = curr_price;
     }
@@ -15,10 +17,10 @@ public class Stock {
     public void setTicker(String ticker) {
         this.ticker = ticker;
     }
-    public Double getCurr_price() {
+    public BigDecimal getCurr_price() {
         return curr_price;
     }
-    public void setCurr_price(Double curr_price) {
+    public void setCurr_price(BigDecimal curr_price) {
         this.curr_price = curr_price;
     }
 }
